+++ .gitignore
@@ -5,4 +5,4 @@ temp/
 node_modules
 .cache
 dist/
-
+.env
