+++ backend_api/cart_checkout.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Cart Checkout</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            margin: 0;
+            padding: 0;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            height: 100vh;
+            background-color: #f4f4f4;
+        }
+        .checkout-message {
+            background-color: #fff;
+            padding: 20px;
+            border-radius: 5px;
+            box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+            text-align: center;
+        }
+        .checkout-message h1 {
+            margin: 0 0 10px;
+            font-size: 24px;
+        }
+        .checkout-message p {
+            margin: 0;
+            font-size: 18px;
+        }
+    </style>
+</head>
+<body>
+    <div class="checkout-message">
+        <h1>Dear, {{name}}</h1>
+        <h2>Thank you for your purchase!</h2>
+        <p>Your order of {{message}} has been successfully placed.</p>
+    </div>
+</body>
+</html>
