+++ database/carts.json
@@ -1,19 +1,7 @@
 [
     {},
-    {},
     {
-        "id": "3ddda2b3-d661-495c-b052-35da78af3b40",
-        "products": [],
-        "total": 0,
-        "discountedTotal": 0,
-        "userId": 1,
-        "totalProducts": 0,
-        "totalQuantity": 0
-    },
-    {},
-    {},
-    {
-        "id": "afd3931a-e5ee-4fac-8981-1b0843ad7ca4",
+        "id": "c68903b2-117d-4f43-ac1e-e1658ed75792",
         "products": [
             {
                 "id": "97d22b8b-b724-4ce1-bf92-b4125a5966be",
@@ -21,42 +9,32 @@
                 "price": 9.99,
                 "discountPercentage": 7.17,
                 "thumbnail": "https://cdn.dummyjson.com/products/images/beauty/Essence%20Mascara%20Lash%20Princess/thumbnail.png",
-                "quantity": 38,
+                "quantity": 19,
                 "total": 9.99,
                 "discountedPrice": 9.273717,
                 "discountedTotal": 9.273717
             },
             {
-                "id": "c92e2bb3-61bc-42b8-9b21-c3eaba19d9f2",
-                "title": "Calvin Klein Heel Shoes",
-                "price": 79.99,
-                "discountPercentage": 11.39,
-                "thumbnail": "https://cdn.dummyjson.com/products/images/womens-shoes/Calvin%20Klein%20Heel%20Shoes/thumbnail.png",
+                "id": "0987eb5a-1b45-4cea-b028-7e58b58f7682",
+                "title": "Gigabyte Aorus Men Tshirt",
+                "price": 24.99,
+                "discountPercentage": 12.6,
+                "thumbnail": "https://cdn.dummyjson.com/products/images/mens-shirts/Gigabyte%20Aorus%20Men%20Tshirt/thumbnail.png",
                 "quantity": 1,
-                "total": 79.99,
-                "discountedPrice": 70.879139,
-                "discountedTotal": 70.879139
+                "total": 24.99,
+                "discountedPrice": 21.84126,
+                "discountedTotal": 21.84126
             },
             {
-                "id": "15c997cb-c1fb-46e3-8798-8b5249e6c8be",
-                "title": "Eyeshadow Palette with Mirror",
-                "price": 19.99,
-                "discountPercentage": 5.5,
-                "thumbnail": "https://cdn.dummyjson.com/products/images/beauty/Eyeshadow%20Palette%20with%20Mirror/thumbnail.png",
-                "quantity": 5,
-                "total": 19.99,
-                "discountedPrice": 18.890549999999998,
-                "discountedTotal": 18.890549999999998
-            }
-        ],
-        "total": 0,
-        "discountedTotal": 0,
-        "userId": 207,
-        "totalProducts": 0,
-        "totalQuantity": 0
-    }
-]        "discountedPrice": 332.000514,
-                "discountedTotal": 332.000514
+                "id": "c837ea6f-26d2-4b8b-8657-f393470aa22c",
+                "title": "Chanel Coco Noir Eau De",
+                "price": 129.99,
+                "discountPercentage": 18.64,
+                "thumbnail": "https://cdn.dummyjson.com/products/images/fragrances/Chanel%20Coco%20Noir%20Eau%20De/thumbnail.png",
+                "quantity": 1,
+                "total": 129.99,
+                "discountedPrice": 105.75986400000001,
+                "discountedTotal": 105.75986400000001
             },
             {
                 "id": "af5bc76c-5046-4635-bf32-e52036679778",
@@ -64,54 +42,66 @@
                 "price": 99.99,
                 "discountPercentage": 16.3,
                 "thumbnail": "https://cdn.dummyjson.com/products/images/mobile-accessories/Amazon%20Echo%20Plus/thumbnail.png",
-                "quantity": 1,
+                "quantity": 2,
                 "total": 99.99,
                 "discountedPrice": 83.69162999999999,
                 "discountedTotal": 83.69162999999999
             },
             {
-                "id": "f42c5240-5fb7-4fad-9bd2-7403abca5ca1",
-                "title": "Rolex Cellini Date Black Dial",
-                "price": 8999.99,
-                "discountPercentage": 14.28,
-                "thumbnail": "https://cdn.dummyjson.com/products/images/mens-watches/Rolex%20Cellini%20Date%20Black%20Dial/thumbnail.png",
-                "quantity": 6,
-                "total": 8999.99,
-                "discountedPrice": 7714.7914279999995,
-                "discountedTotal": 7714.7914279999995
-            },
-            {
-                "id": "97d22b8b-b724-4ce1-bf92-b4125a5966be",
-                "title": "Essence Mascara Lash Princess",
-                "price": 9.99,
-                "discountPercentage": 7.17,
-                "thumbnail": "https://cdn.dummyjson.com/products/images/beauty/Essence%20Mascara%20Lash%20Princess/thumbnail.png",
+                "id": "15c997cb-c1fb-46e3-8798-8b5249e6c8be",
+                "title": "Eyeshadow Palette with Mirror",
+                "price": 19.99,
+                "discountPercentage": 5.5,
+                "thumbnail": "https://cdn.dummyjson.com/products/images/beauty/Eyeshadow%20Palette%20with%20Mirror/thumbnail.png",
                 "quantity": 2,
-                "total": 9.99,
-                "discountedPrice": 9.273717,
-                "discountedTotal": 9.273717
+                "total": 19.99,
+                "discountedPrice": 18.890549999999998,
+                "discountedTotal": 18.890549999999998
             },
             {
-                "id": "ccab5bc8-9425-4912-995a-3ce9a1705037",
-                "title": "iPhone 5s",
-                "price": 199.99,
-                "discountPercentage": 11.85,
-                "thumbnail": "https://cdn.dummyjson.com/products/images/smartphones/iPhone%205s/thumbnail.png",
-                "quantity": 1,
-                "total": 199.99,
-                "discountedPrice": 176.29118499999998,
-                "discountedTotal": 176.29118499999998
+                "id": "64394358-ef92-4013-8902-566c499e9ff3",
+                "title": "iPhone 13 Pro",
+                "price": 1099.99,
+                "discountPercentage": 18.3,
+                "thumbnail": "https://cdn.dummyjson.com/products/images/smartphones/iPhone%2013%20Pro/thumbnail.png",
+                "quantity": 4,
+                "total": 1099.99,
+                "discountedPrice": 898.69183,
+                "discountedTotal": 898.69183
+            },
+            {
+                "id": "5155bb21-ea6d-421b-a4e7-edbca52e6493",
+                "title": "iPhone X",
+                "price": 899.99,
+                "discountPercentage": 14.19,
+                "thumbnail": "https://cdn.dummyjson.com/products/images/smartphones/iPhone%20X/thumbnail.png",
+                "quantity": 2,
+                "total": 899.99,
+                "discountedPrice": 772.281419,
+                "discountedTotal": 772.281419
             }
         ],
-        "total": 0.0,
-        "discountedTotal": 0.0,
+        "total": 0,
+        "discountedTotal": 0,
         "userId": 207,
         "totalProducts": 0,
         "totalQuantity": 0
     },
     {
-        "id": "3ddda2b3-d661-495c-b052-35da78af3b40",
-        "products": [],
+        "id": "a04d8b9e-e437-4638-b9ef-c90e69e8ee86",
+        "products": [
+            {
+                "id": "f3108b1f-36db-42c6-b0e8-972e81bbf3c3",
+                "title": "Watch Gold for Women",
+                "price": 799.99,
+                "discountPercentage": 6.37,
+                "thumbnail": "https://cdn.dummyjson.com/products/images/womens-watches/Watch%20Gold%20for%20Women/thumbnail.png",
+                "quantity": 6,
+                "total": 799.99,
+                "discountedPrice": 749.0306370000001,
+                "discountedTotal": 749.0306370000001
+            }
+        ],
         "total": 0,
         "discountedTotal": 0,
         "userId": 1,
