+++ posts.html
@@ -6,6 +6,7 @@
     <title>Post</title>
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
     <script src="js/scripts.js"></script>
+    <script src="js/posts.js"></script>
     <link
             href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
             rel="stylesheet"
