+++ product.html
@@ -87,6 +87,7 @@
             </div>
         </div>
         <script src = "js/scripts.js"></script>
+        <script src = "js/product.js"></script>
         <script>
             loadProductPage()
         </script>
