+++ search.html
@@ -8,6 +8,8 @@
     <!-- Bootstrap CSS -->
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
     <script src="js/scripts.js"></script>
+    <script src = "js/search.js"></script>
+    
     <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
     <style>
         body {
